Wallet: Separate fake agora, stoa and sample data needed for testing into separate files
